LPS-172038 portal-search-web: pull up testOneTermWithPreviousSelection
https://issues.liferay.com/browse/LPS-172038O
