+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/AssetEntriesSearchFacetDisplayContextTest.java
@@ -54,6 +54,16 @@ public class AssetEntriesSearchFacetDisplayContextTest
 
 	@Override
 	@Test
+	public void testOneTerm() throws Exception {
+	}
+
+	@Override
+	@Test
+	public void testOneTermWithPreviousSelection() throws Exception {
+	}
+
+	@Override
+	@Test
 	public void testOrderByTermValueAscending() throws Exception {
 		testOrderBy(
 			new String[] {"bravo", "alpha", "charlie"}, new int[] {2, 3, 4},
