+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/BaseCategoriesSearchFacetDisplayContextTestCase.java
@@ -187,6 +187,7 @@ public abstract class BaseCategoriesSearchFacetDisplayContextTestCase
 		Assert.assertFalse(facetDisplayContext.isRenderNothing());
 	}
 
+	@Override
 	@Test
 	public void testOneTermWithPreviousSelection() throws Exception {
 		long assetCategoryId = RandomTestUtil.randomLong();
