+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/ScopeSearchFacetDisplayContextTest.java
@@ -26,14 +26,11 @@ import com.liferay.portal.search.web.internal.site.facet.configuration.SiteFacet
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 
-import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
-import org.junit.Test;
 
 import org.mockito.Mockito;
 
@@ -79,47 +76,6 @@ public class ScopeSearchFacetDisplayContextTest
 	@Override
 	public String getFacetDisplayContextParameterValue() {
 		return "0";
-	}
-
-	@Test
-	public void testOneTermWithPreviousSelection() throws Exception {
-		long groupId = RandomTestUtil.randomLong();
-		String name = RandomTestUtil.randomString();
-
-		_addGroup(groupId, name);
-
-		int frequency = RandomTestUtil.randomInt();
-
-		setUpTermCollectors(
-			facetCollector,
-			Collections.singletonList(
-				createTermCollector(String.valueOf(groupId), frequency)));
-
-		String parameterValue = String.valueOf(groupId);
-
-		FacetDisplayContext facetDisplayContext = createFacetDisplayContext(
-			parameterValue);
-
-		List<BucketDisplayContext> bucketDisplayContexts =
-			facetDisplayContext.getBucketDisplayContexts();
-
-		Assert.assertEquals(
-			bucketDisplayContexts.toString(), 1, bucketDisplayContexts.size());
-
-		BucketDisplayContext bucketDisplayContext = bucketDisplayContexts.get(
-			0);
-
-		Assert.assertEquals(name, bucketDisplayContext.getBucketText());
-		Assert.assertEquals(
-			String.valueOf(groupId), bucketDisplayContext.getFilterValue());
-		Assert.assertEquals(frequency, bucketDisplayContext.getFrequency());
-		Assert.assertTrue(bucketDisplayContext.isFrequencyVisible());
-		Assert.assertTrue(bucketDisplayContext.isSelected());
-
-		Assert.assertEquals(
-			parameterValue, facetDisplayContext.getParameterValue());
-		Assert.assertFalse(facetDisplayContext.isNothingSelected());
-		Assert.assertFalse(facetDisplayContext.isRenderNothing());
 	}
 
 	protected Group createGroup(long groupId, String name) throws Exception {
